Implement MAUI Client UI (WIP)
-Implement MAUI client login and create transaction behaviour (WIP)
-Move MAUI client work to broken project to migrate to working project
-Create new going-to-actually-work MAUI client
-Implement error handling for server getting invalid request

+++ FinanceApp.Client/Classes/FinanceClient.cs
@@ -72,7 +72,7 @@ public class FinanceClient : IClient
 					{
 						Create transaction = new()
 						{
-                            Value = new RequestField<long>
+							Value = new RequestField<double>
                             {
                                 Value = value
                             },
