+++ FinanceApp.Data/Models/Transaction.cs
@@ -4,7 +4,7 @@ public class Transaction
 {
     public long Id { get; set; }
     public required Counterparty Counterparty { get; set; }
-    public required long Value { get; set; }
+    public required double Value { get; set; }
 
     public override string ToString()
     {
