+++ FinanceApp.Data/Requests/Transaction/Create.cs
@@ -8,7 +8,7 @@ public class Create : ISingleTransaction
 {
     public static string Flag => "<CreateTransaction>";
 
-    public required RequestField<long> Value { get; init; }
+    public required RequestField<double> Value { get; init; }
     public required RequestField<Counterparty> Counterparty { get; init; }
 
     public override string ToString()
